Finished GroupLevel.cpp .
Need to do final cleanup and remove excess functions.
Added a hook into the group update timer to check the players in a group on that tick.
Additions to ScriptMgr.h and ScriptMgr.cpp .
Made adjustments to the bruteforce skill learn for """"racials""""
Added sprint autolearn and its config toggle to Player::LearnDefaultSkills instead of AutoSpells.cpp .